+++ apps/boruta_identity/priv/repo/migrations/20221028083326_add_revoked_at_to_users_tokens.exs
@@ -1,0 +1,9 @@
+defmodule BorutaIdentity.Repo.Migrations.AddRevokedAtToUsersTokens do
+  use Ecto.Migration
+
+  def change do
+    alter table(:users_tokens) do
+      add :revoked_at, :utc_datetime_usec
+    end
+  end
+end